Finished refactor, now for tests :)
Changed filename to better reflect static behavior.
